% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafico de temperatura mensual}
\usage{
grafico_temperatura_mensual(
  df,
  colores = NULL,
  titulo = "Temperatura promedio"
)
}
\arguments{
\item{df}{Data frame con las columnas id, fecha y temperatura_abrigo_150cm.}

\item{colores}{Vector opcional de colores.}

\item{titulo}{Titulo del grafico.}
}
\value{
Un grafico de ggplot2.
}
\description{
Muestra la temperatura promedio mensual de una o mas estaciones.
}
\examples{
\dontrun{
df <- leer_datos_estacion("NH0437", "NH0437.csv")
grafico_temperatura_mensual(df)
}

}
