% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico de temperatura mensual}
\usage{
grafico_temperatura_mensual(
  df,
  colores = NULL,
  titulo = "Temperatura promedio"
)
}
\arguments{
\item{df}{Data frame con las columnas id, fecha y temperatura_abrigo_150cm.}

\item{colores}{Vector opcional de colores.}

\item{titulo}{Título del gráfico.}
}
\value{
Un gráfico de ggplot2.
}
\description{
Muestra la temperatura promedio mensual de una o más estaciones.
}
