% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Generar una tabla resumen de temperatura}
\usage{
tabla_resumen_temperatura(...)
}
\arguments{
\item{...}{Uno o más data frames devueltos por \code{leer_datos_estacion()}}
}
\value{
Un data frame con los valores resumen de temperatura
}
\description{
Calcula, para cada estación, la media, mínimo, máximo y desviación
estándar de la temperatura a 150 cm de abrigo.
}
